refactor: implement XLSX file reading functionality using ExcelJS in user-management
